remove container only if exists

+++ bin/before-install.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-# stop and remove container
-docker stop home-credit-risk
-docker rm home-credit-risk
+if docker ps | grep " $PROJECT_NAME$"; then
+    # stop and remove container
+    docker stop home-credit-risk
+    docker rm home-credit-risk
 
-# delete image
-docker rmi home-credit-risk
+    # delete image
+    docker rmi home-credit-risk
+fi
